Refactored classes (by adding packages) and made Core, which seems to be an attempt at a generic plugin base, such.
